Move relays to separate folders.
