+++ piksi_multi_cpp/CMakeLists.txt
@@ -21,9 +21,9 @@ cs_add_library(${PROJECT_NAME}
   src/receiver/receiver.cc
   src/sbp_callback_handler/sbp_callback_handler.cc
   src/sbp_callback_handler/sbp_callback_handler_factory.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay_heartbeat.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay_imu_raw.cc
+  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay.cc
+  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_heartbeat.cc
+  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_imu_raw.cc
 )
 target_link_libraries(${PROJECT_NAME} serialport)
