+++ piksi_multi_cpp/include/piksi_multi_cpp/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_heartbeat.h
@@ -3,7 +3,7 @@
 
 #include <libsbp/system.h>
 #include <piksi_rtk_msgs/Heartbeat.h>
-#include "piksi_multi_cpp/sbp_callback_handler/sbp_callback_handler_relay.h"
+#include "piksi_multi_cpp/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay.h"
 
 namespace piksi_multi_cpp {
