--- piksi_multi_cpp/src/sbp_callback_handler/sbp_callback_handler_relay.cc
@@ -1,1 +1,0 @@
-#include "piksi_multi_cpp/sbp_callback_handler/sbp_callback_handler_relay.h"
